packages/django-rest/requirements.txt
asgiref==3.8.1
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
channels==4.0.0
channels-redis==4.2.0
daphne
dj_database_url==2.1.0
drf-extra-fields==3.7.0
gunicorn==22.0.0
pillow==10.3.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
whitenoise==6.6.0
requests
django-filter